chore(agents-api): moved tool_executor model to autogen
